pyproject.toml
[tool.poetry]
name = "ideation"
version = "0.1.0"
description = "Ideation workspace automation"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {version = "^0.29.0", extras = ["standard"]}
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
greenlet = "^3.0.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
honcho = "^2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
